Load BdE full time-series files

Load a full time-series file provided by BdE.

bde_series_full_load(
  series_csv,
  parse_dates = TRUE,
  parse_numeric = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  extract_metadata = FALSE
)

Arguments

series_csv

csv file of a series, as defined in the field Nombre del archivo con los valores de la serie of the corresponding catalog. See bde_catalog_load().

parse_dates

Logical. If TRUE the dates would be parsed using bde_parse_dates().

parse_numeric

Logical. If TRUE the columns would be parsed to double (numeric) values. See Note.

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache

Logical. If TRUE the requested file would be updated on the cache_dir.

verbose

Logical TRUE or FALSE, display information useful for debugging.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is the metadata of the requested series.

Value

A tibble with a field "Date" and the alias of the fields series as described on the catalogs. See bde_catalog_load().

Details

About BdE file naming

The series name is a positional code showing the location of the table. For example, table be_6_1 represents the Table 1, Chapter 6 of the Statistical Bulletin ("BE"). Although it is a unique value, it is subject to change (i.e. a new table is inserted before).

For that reason, the function bde_series_load() is more suitable for extracting specific time-series.

Note

This function tries to coerce the columns to numbers. For some series a warning may be displayed if the parser fails. You can override the default behavior with parse_numeric = FALSE

See also

Examples

# \donttest{
# Metadata
bde_series_full_load("TI_1_1.csv", extract_metadata = TRUE)
#> # A tibble: 6 x 9
#>   Date  TI_1_1.1 TI_1_1.2 TI_1_1.3 TI_1_1.4 TI_1_1.5 TI_1_1.7 TI_1_1.9 TI_1_1.10
#>   <chr> <chr>    <chr>    <chr>    <chr>    <chr>    <chr>    <chr>    <chr>    
#> 1 NOMB~ D_DTFK0~ D_DTFK0~ D_DTFK0~ D_DNBBB~ D_DBEI0~ D_DBEA0~ D_DNBBB~ D_DNBBBB~
#> 2 NÚME~ 1654131  1654132  1654133  903705   1589846  1589850  581532   581533   
#> 3 ALIA~ TI_1_1.1 TI_1_1.2 TI_1_1.3 TI_1_1.4 TI_1_1.5 TI_1_1.7 TI_1_1.9 TI_1_1.10
#> 4 DESC~ Tipo de~ Tipo de~ Tipo de~ Tipo de~ Tipo de~ Tipo de~ Tipo de~ Tipo de ~
#> 5 DESC~ Porcent~ Céntimo~ Porcent~ Porcent~ Porcent~ Porcent~ Porcent~ Porcenta~
#> 6 FREC~ DIARIA   DIARIA   DIARIA   DIARIA   DIARIA   DIARIA   DIARIA   DIARIA   

# Data
bde_series_full_load("TI_1_1.csv")
#> # A tibble: 11,539 x 9
#>    Date       TI_1_1.1 TI_1_1.2 TI_1_1.3 TI_1_1.4 TI_1_1.5 TI_1_1.7 TI_1_1.9
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 1990-05-14       NA     14.6     14.6       NA       NA       NA       NA
#>  2 1990-05-15       NA     NA       NA         NA       NA       NA       NA
#>  3 1990-05-16       NA     NA       NA         NA       NA       NA       NA
#>  4 1990-05-17       NA     NA       NA         NA       NA       NA       NA
#>  5 1990-05-18       NA     NA       NA         NA       NA       NA       NA
#>  6 1990-05-19       NA     NA       NA         NA       NA       NA       NA
#>  7 1990-05-20       NA     NA       NA         NA       NA       NA       NA
#>  8 1990-05-21       NA     NA       NA         NA       NA       NA       NA
#>  9 1990-05-22       NA     NA       NA         NA       NA       NA       NA
#> 10 1990-05-23       NA     14.6     14.6       NA       NA       NA       NA
#> # ... with 11,529 more rows, and 1 more variable: TI_1_1.10 <dbl>
# }

Site built with pkgdown 2.0.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.