Load BdE catalogs

Load the time-series catalogs provided by BdE.

bde_catalog_load(
  catalog = "ALL",
  parse_dates = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)

Source

Time-series bulk data download

Arguments

catalog

A single value indicating the catalogs to be updated or "ALL" as a shorthand. See Details.

parse_dates

Logical. If TRUE the dates would be parsed using bde_parse_dates().

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache

Logical. If TRUE the requested file would be updated on the cache_dir.

verbose

Logical TRUE or FALSE, display information useful for debugging.

Value

A tibble

Details

Accepted values for catalog are:

CODE	PUBLICATION	UPDATE FREQUENCY	FREQUENCY

BE	Statistical Bulletin	Daily	Monthly

CF	Financial Accounts of the Spanish Economy	Quarterly	Annual

IE (see Note)	Economic Indicators	Daily	Monthly

SI	Summary Indicators	Daily	Daily

TC	Exchange Rates	Daily	Daily

TI	Interest Rates	Daily	Daily

PB	Bank Lending Survey	Quarterly	Quarterly

Use "ALL" as a shorthand for updating all the catalogs at a glance.

If the requested catalog is not cached bde_catalog_update() is invoked.

Note

From 1 July 2021, the contents of the publication Economic Indicators will gradually be migrated to the Statistical Bulletin. From June 2022, the Economic Indicators section will no longer be available.

See also

Examples

# \donttest{
bde_catalog_load("TI", verbose = TRUE)
#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\RtmpQdILnr
#> tidyBdE> Cache dir is C:\Users\RUNNER~1\AppData\Local\Temp\RtmpQdILnr
#> tidyBdE> Updating catalogs: TI
#> tidyBdE> Downloading file from https://www.bde.es/webbde/es/estadis/infoest/series/catalogo_ti.csv
#> 
#> tidyBdE> Parsing dates
#> # A tibble: 68 x 17
#>    Nombre_de_la_serie Numero_secuencial Alias_de_la_serie Nombre_del_archivo_co~
#>    <chr>                          <dbl> <chr>             <chr>                 
#>  1 D_DTFK01A0                   1654131 TI_1_1.1          TI_1_1.csv            
#>  2 D_DTFK03A0                   1654132 TI_1_1.2          TI_1_1.csv            
#>  3 D_DTFK02A0                   1654133 TI_1_1.3          TI_1_1.csv            
#>  4 D_DNBBBP72                    903705 TI_1_1.4          TI_1_1.csv            
#>  5 D_DBEI00AJ00F                1589846 TI_1_1.5          TI_1_1.csv            
#>  6 D_DBEA00AJ00F                1589850 TI_1_1.7          TI_1_1.csv            
#>  7 D_DNBBBA72                    581532 TI_1_1.9          TI_1_1.csv            
#>  8 D_DNBBBB72                    581533 TI_1_1.10         TI_1_1.csv            
#>  9 D_DE721006                   1654134 TI_1_2.1          TI_1_2.csv            
#> 10 D_DE722706                     58233 TI_1_2.2          TI_1_2.csv            
#> # ... with 58 more rows, and 13 more variables: Descripcion_de_la_serie <chr>,
#> #   Tipo_de_variable <chr>, Codigo_de_unidades <chr>, Exponente <dbl>,
#> #   Numero_de_decimales <dbl>, Descripcion_de_unidades_y_exponente <chr>,
#> #   Frecuencia_de_la_serie <chr>, Fecha_de_la_primera_observacion <date>,
#> #   Fecha_de_la_ultima_observacion <date>, Numero_de_observaciones <dbl>,
#> #   Titulo_de_la_serie <chr>, Fuente <chr>, Notas <chr>
# }

Site built with pkgdown 2.0.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.