Load BdE catalogs

Load the time-series catalogs provided by BdE.

Usage

bde_catalog_load(
  catalog = "ALL",
  parse_dates = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)

Source

Time-series bulk data download

Arguments

catalog

A single value indicating the catalogs to be updated or "ALL" as a shorthand. See Details.

parse_dates

Logical. If TRUE the dates would be parsed using bde_parse_dates().

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache

Logical. If TRUE the requested file would be updated on the cache_dir.

verbose

Logical TRUE or FALSE, display information useful for debugging.

Value

A tibble

Details

Accepted values for catalog are:

CODE	PUBLICATION	UPDATE FREQUENCY	FREQUENCY

BE	Statistical Bulletin	Daily	Monthly

SI	Summary Indicators	Daily	Daily

TC	Exchange Rates	Daily	Daily

TI	Interest Rates	Daily	Daily

PB	Bank Lending Survey	Quarterly	Quarterly

Use "ALL" as a shorthand for updating all the catalogs at a glance.

If the requested catalog is not cached bde_catalog_update() is invoked.

See also

Examples

# \donttest{
bde_catalog_load("TI", verbose = TRUE)
#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\RtmpGEsG8r
#> tidyBdE> Need to download catalog TI
#> tidyBdE> Cache dir is C:\Users\RUNNER~1\AppData\Local\Temp\RtmpGEsG8r
#> tidyBdE> Updating catalogs: TI
#> tidyBdE> Downloading file from https://www.bde.es/webbde/es/estadis/infoest/series/catalogo_ti.csv
#> 
#> tidyBdE> Parsing dates
#> # A tibble: 68 × 17
#>    Nombre_de_l…¹ Numer…² Alias…³ Nombr…⁴ Descr…⁵ Tipo_…⁶ Codig…⁷ Expon…⁸ Numer…⁹
#>    <chr>           <dbl> <chr>   <chr>   <chr>   <chr>   <chr>     <dbl>   <dbl>
#>  1 D_DTFK01A0    1654131 TI_1_1… TI_1_1… "Tipo … MEDIA   Porcen…      -2       3
#>  2 D_DTFK03A0    1654132 TI_1_1… TI_1_1… "Tipo … MEDIA   EUR          -2       3
#>  3 D_DTFK02A0    1654133 TI_1_1… TI_1_1… "Tipo … MEDIA   Porcen…      -2       3
#>  4 D_DNBBBP72     903705 TI_1_1… TI_1_1… "Tipo … FINAL   Porcen…      -2       2
#>  5 D_DBEI00AJ00F 1589846 TI_1_1… TI_1_1… "Tipo … MEDIA   Porcen…      -2       3
#>  6 D_DBEA00AJ00F 1589850 TI_1_1… TI_1_1… "Tipo … MEDIA   Porcen…      -2       3
#>  7 D_DNBBBA72     581532 TI_1_1… TI_1_1… "Tipo … FINAL   Porcen…      -2       2
#>  8 D_DNBBBB72     581533 TI_1_1… TI_1_1… "Tipo … FINAL   Porcen…      -2       2
#>  9 D_DE721006    1654134 TI_1_2… TI_1_2… "Merca… MEDIA   Porcen…      -2       3
#> 10 D_DE722706      58233 TI_1_2… TI_1_2…  NA     MEDIA   Porcen…      -2       3
#> # … with 58 more rows, 8 more variables:
#> #   Descripcion_de_unidades_y_exponente <chr>, Frecuencia_de_la_serie <chr>,
#> #   Fecha_de_la_primera_observacion <date>,
#> #   Fecha_de_la_ultima_observacion <date>, Numero_de_observaciones <dbl>,
#> #   Titulo_de_la_serie <chr>, Fuente <chr>, Notas <chr>, and abbreviated
#> #   variable names ¹​Nombre_de_la_serie, ²​Numero_secuencial, ³​Alias_de_la_serie,
#> #   ⁴​Nombre_del_archivo_con_los_valores_de_la_serie, …
# }

Site built with pkgdown 2.0.6.

Template rostemplate by dieghernan, based on Bootstrapious.