Update BdE catalogs

Update the time-series catalogs provided by BdE.

bde_catalog_update(catalog = "ALL", cache_dir = NULL, verbose = FALSE)

Source

Time-series bulk data download

Arguments

catalog

A vector of characters indicating the catalogs to be updated or "ALL" as a shorthand. See Details.

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

verbose

Logical TRUE or FALSE, display information useful for debugging.

Value

None Downloads the catalog file(s) to the local machine.

Details

Accepted values for catalog are:

CODE	PUBLICATION	UPDATE FREQUENCY	FREQUENCY

BE	Statistical Bulletin	Daily	Monthly

CF	Financial Accounts of the Spanish Economy	Quarterly	Annual

IE	Economic Indicators	Daily	Monthly

SI	Summary Indicators	Daily	Daily

TC	Exchange Rates	Daily	Daily

TI	Interest Rates	Daily	Daily

PB	Bank Lending Survey	Quarterly	Quarterly

Use "ALL" as a shorthand for updating all the catalogs at a glance.

See also

Examples

# \donttest{
bde_catalog_update("TI", verbose = TRUE)
#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\RtmpQdILnr
#> tidyBdE> Updating catalogs: TI
#> tidyBdE> Downloading file from https://www.bde.es/webbde/es/estadis/infoest/series/catalogo_ti.csv
#> 
#> [1] TRUE
# }

Site built with pkgdown 2.0.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.