Relevant Indicators of Spain

Set of helper functions for downloading some of the most relevant macroeconomic indicators of Spain.

bde_ind_gdp_var(series_label = "GDP_YoY", ...)

bde_ind_unemployment_rate(series_label = "Unemployment_Rate", ...)

bde_ind_euribor_12m_monthly(series_label = "Euribor_12M_Monthly", ...)

bde_ind_euribor_12m_daily(series_label = "Euribor_12M_Daily", ...)

bde_ind_cpi_var(series_label = "Consumer_price_index_YoY", ...)

bde_ind_ibex(series_label = "IBEX_index_month", ...)

Arguments

series_label
Optional. Character vector or value. Allows to specify a custom label for the series extracted. It should have the same length than series_code.

...
Arguments passed on to bde_series_load

parse_dates

Logical. If TRUE the dates would be parsed using bde_parse_dates().

parse_numeric

Logical. If TRUE the columns would be parsed to double (numeric) values. See Note.

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache

Logical. If TRUE the requested file would be updated on the cache_dir.

verbose

Logical, display information useful for debugging.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is the metadata of the requested series.

Value

A tibble with the required series.

Details

This functions are convenient wrappers of bde_series_load() referencing specific series. Use verbose = TRUE/extract_metadata = TRUE to see the specification and the source.

See also

Examples

# \donttest{
bde_ind_euribor_12m_monthly()

#> # A tibble: 245 x 2
#>    Date       Euribor_12M_Monthly
#>    <date>                   <dbl>
#>  1 2001-01-01                4.57
#>  2 2001-02-01                4.59
#>  3 2001-03-01                4.47
#>  4 2001-04-01                4.48
#>  5 2001-05-01                4.52
#>  6 2001-06-01                4.31
#>  7 2001-07-01                4.31
#>  8 2001-08-01                4.11
#>  9 2001-09-01                3.77
#> 10 2001-10-01                3.37
#> # … with 235 more rows

# }

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.