Relevant Indicators of Spain

Set of helper functions for downloading some of the most relevant macroeconomic indicators of Spain.

Usage

bde_ind_gdp_var(series_label = "GDP_YoY", ...)

bde_ind_unemployment_rate(series_label = "Unemployment_Rate", ...)

bde_ind_euribor_12m_monthly(series_label = "Euribor_12M_Monthly", ...)

bde_ind_euribor_12m_daily(series_label = "Euribor_12M_Daily", ...)

bde_ind_cpi_var(series_label = "Consumer_price_index_YoY", ...)

bde_ind_ibex(series_label = "IBEX_index_month", ...)

bde_ind_gdp_quarterly(series_label = "GDP_quarterly_value", ...)

bde_ind_population(series_label = "Population_Spain", ...)

Arguments

series_label

Optional. Character vector or value. Allows to specify a custom label for the series extracted. It should have the same length than series_code.

...

Arguments passed on to bde_series_load

parse_dates

Logical. If TRUE the dates would be parsed using bde_parse_dates().

parse_numeric

Logical. If TRUE the columns would be parsed to double (numeric) values. See Note.

cache_dir

A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache

Logical. If TRUE the requested file would be updated on the cache_dir.

verbose

Logical TRUE or FALSE, display information useful for debugging.

extract_metadata

Logical TRUE/FALSE. On TRUE the output is the metadata of the requested series.

Value

A tibble with the required series.

Details

This functions are convenient wrappers of bde_series_load() referencing specific series. Use verbose = TRUE, extract_metadata = TRUE options to see the specification and the source.

See also

Examples

# \donttest{
bde_ind_gdp_var()
#> # A tibble: 105 × 2
#>    Date       GDP_YoY
#>    <date>       <dbl>
#>  1 1996-03-01    2.4 
#>  2 1996-06-01    2.52
#>  3 1996-09-01    2.98
#>  4 1996-12-01    2.75
#>  5 1997-03-01    3.18
#>  6 1997-06-01    3.35
#>  7 1997-09-01    3.69
#>  8 1997-12-01    4.57
#>  9 1998-03-01    4.42
#> 10 1998-06-01    4.65
#> # … with 95 more rows
# }

Site built with pkgdown 2.0.3.

Template rostemplate by dieghernan, based on Bootstrapious.