Load a single BdE time-series.

The series alias is a positional code showing the location (column and/or row) of the series in the table. However, although it is unique, it is not a good candidate to be used as the series ID, as it is subject to change. If a series changes position in the table, its alias will also change.

To ensure series can still be identified, even after these changes, they are assigned a sequential number (series_code on this function) which will remain unchanged throughout the series' lifetime.

Note that a single series could be used on different tables, so it can have several aliases. If you need to search by alias it is recommended to use bde_series_full_load().

bde_series_load(
  series_code,
  series_label = NULL,
  parse_dates = TRUE,
  parse_numeric = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  extract_metadata = FALSE
)

Arguments

series_code
a numeric value or vector with time-series code(s), as defined in the field Número secuencial of the corresponding series. See bde_catalog_load().

series_label
Optional. Character vector or value. Allows to specify a custom label for the series extracted. It should have the same length than series_code.

parse_dates
Logical. If TRUE the dates would be parsed using bde_parse_dates().

parse_numeric
Logical. If TRUE the columns would be parsed to double (numeric) values. See Note.

cache_dir
A path to a cache directory. The directory can also be set via options with options(bde_cache_dir = "path/to/dir").

update_cache
Logical. If TRUE the requested file would be updated on the cache_dir.

verbose
Logical, display information useful for debugging.

extract_metadata
Logical TRUE/FALSE. On TRUE the output is the metadata of the requested series.

Value

A tibble with a field "Date" and the series_code of the series as described on the catalogs. See bde_catalog_load().

Details

Load a single time-series provided by BdE.

Note

This function tries to coerce the columns to numbers. For some series a warning may be displayed if the parser fails. You can override the default behavior with parse_numeric = FALSE

See also

Examples

# \donttest{
# Metadata
bde_series_load(573234, verbose = TRUE, extract_metadata = TRUE)

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K

#> tidyBdE> Parsing dates

#> tidyBdE> Extracting series 573234
#> 

#> tidyBdE> Downloading serie 573234 from file TC_1_1.csv (alias TC_1_1.1).

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K/TC

#> tidyBdE> Downloading file from https://www.bde.es/webbde/es/estadis/infoest/series/tc_1_1.csv
#> 

#> # A tibble: 6 x 2
#>   Date                        `573234`                                          
#>   <chr>                       <chr>                                             
#> 1 NOMBRE DE LA SERIE          DTCCBCEUSDEUR.B                                   
#> 2 NÚMERO SECUENCIAL           573234                                            
#> 3 ALIAS DE LA SERIE           TC_1_1.1                                          
#> 4 DESCRIPCIÓN DE LA SERIE     Tipo de cambio. Dólares estadounidenses por euro ~
#> 5 DESCRIPCIÓN DE LAS UNIDADES Dólares de Estados Unidos por Euro                
#> 6 FRECUENCIA                  LABORABLE                                         

# Data
bde_series_load(573234, extract_metadata = FALSE)

#> # A tibble: 5,892 x 2
#>    Date       `573234`
#>    <date>        <dbl>
#>  1 1999-01-04     1.18
#>  2 1999-01-05     1.18
#>  3 1999-01-06     1.17
#>  4 1999-01-07     1.16
#>  5 1999-01-08     1.17
#>  6 1999-01-11     1.16
#>  7 1999-01-12     1.15
#>  8 1999-01-13     1.17
#>  9 1999-01-14     1.17
#> 10 1999-01-15     1.16
#> # ... with 5,882 more rows

# Vectorized
bde_series_load(c(573234, 573214),
  series_label = c("US/EUR", "GBP/EUR"),
  verbose = TRUE, extract_metadata = TRUE
)

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K

#> tidyBdE> Parsing dates

#> tidyBdE> Extracting series 573234
#> 

#> tidyBdE> Downloading serie 573234 from file TC_1_1.csv (alias TC_1_1.1).

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K/TC

#> tidyBde> Reading file tc_1_1.csv from cache.

#> tidyBdE> Extracting series 573214
#> 

#> tidyBdE> Downloading serie 573214 from file TC_1_1.csv (alias TC_1_1.4).

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K/TC

#> tidyBde> Reading file tc_1_1.csv from cache.

#> # A tibble: 6 x 3
#>   Date                        `US/EUR`                   `GBP/EUR`              
#>   <chr>                       <chr>                      <chr>                  
#> 1 ALIAS DE LA SERIE           TC_1_1.1                   TC_1_1.4               
#> 2 DESCRIPCIÓN DE LA SERIE     Tipo de cambio. Dólares e~ Tipo de cambio. Libras~
#> 3 DESCRIPCIÓN DE LAS UNIDADES Dólares de Estados Unidos~ Libras esterlinas por ~
#> 4 FRECUENCIA                  LABORABLE                  LABORABLE              
#> 5 NOMBRE DE LA SERIE          DTCCBCEUSDEUR.B            DTCCBCEGBPEUR.B        
#> 6 NÚMERO SECUENCIAL           573234                     573214                 

bde_series_load(c(573234, 573214),
  series_label = c("US/EUR", "GBP/EUR"),
  verbose = TRUE, extract_metadata = FALSE
)

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K

#> tidyBdE> Parsing dates

#> tidyBdE> Extracting series 573234
#> 

#> tidyBdE> Downloading serie 573234 from file TC_1_1.csv (alias TC_1_1.1).

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K/TC

#> tidyBde> Reading file tc_1_1.csv from cache.

#> tidyBdE> Parsing dates

#> tidyBdE> Parsing fields to double

#> tidyBdE> Extracting series 573214
#> 

#> tidyBdE> Downloading serie 573214 from file TC_1_1.csv (alias TC_1_1.4).

#> tidyBdE> Caching on temporary directory C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpqg8r6K/TC

#> tidyBde> Reading file tc_1_1.csv from cache.

#> tidyBdE> Parsing dates

#> tidyBdE> Parsing fields to double

#> # A tibble: 5,892 x 3
#>    Date       `US/EUR` `GBP/EUR`
#>    <date>        <dbl>     <dbl>
#>  1 1999-01-04     1.18     0.711
#>  2 1999-01-05     1.18     0.712
#>  3 1999-01-06     1.17     0.708
#>  4 1999-01-07     1.16     0.706
#>  5 1999-01-08     1.17     0.709
#>  6 1999-01-11     1.16     0.704
#>  7 1999-01-12     1.15     0.707
#>  8 1999-01-13     1.17     0.708
#>  9 1999-01-14     1.17     0.706
#> 10 1999-01-15     1.16     0.704
#> # ... with 5,882 more rows

# }

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.